requirement.txt
Flask==2.3.3
Werkzeug==2.3.7
dnspython==2.4.2
requests==2.31.0
tldextract==3.4.4
python-dotenv==1.0.0
